게임 컴포넌트 레이아웃 변경
